requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lyricsgenius==3.0.1
openai==1.13.3
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-telegram-bot==21.3
redis==5.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spotipy==2.23.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
